update service methods to take role parameters

+++ back-end/service/recipe.service.ts
@@ -1,22 +1,39 @@
+import { UnauthorizedError } from 'express-jwt';
 import { Recipe } from '../model/recipe';
 import recipeDb from '../repository/recipe.db';
-import { RecipeUpdateInput } from '../types';
+import { RecipeUpdateInput, Role } from '../types';
 
-const getAllRecipes = async (): Promise<Recipe[]> => {
-    return await recipeDb.getAllRecipes();
+const getAllRecipes = async (userId: number, role: Role): Promise<Recipe[]> => {
+    if (role === 'admin') {
+        return await recipeDb.getAllRecipes();
+    } else if (role === 'user') {
+        return await recipeDb.getRecipesByUserId(userId);
+    } else {
+        throw new UnauthorizedError('credentials_required', {
+            message: 'Guests can only see recipes from invited users.',
+        });
+    }
 };
 
 const getRecipeById = async (id: number): Promise<Recipe> => {
     const recipe = await recipeDb.getRecipeById({ id });
     if (!recipe) throw new Error(`Recipe with id ${id} does not exist.`);
+
     return recipe;
 };
 
 const updateRecipe = async (
     id: number,
     recipeData: RecipeUpdateInput,
-    userId: number
+    userId: number,
+    role: Role
 ): Promise<Recipe> => {
+    if (role !== 'user' && role !== 'admin') {
+        throw new UnauthorizedError('credentials_required', {
+            message: 'Only users can update their own recipes.',
+        });
+    }
+
     const recipe = await recipeDb.getRecipeById({ id });
     if (!recipe) throw new Error(`Recipe with id ${id} does not exist.`);
 
@@ -29,11 +46,18 @@ const updateRecipe = async (
     return recipe;
 };
 
-const deleteRecipe = async (id: number): Promise<void> => {
+const deleteRecipe = async (id: number, role: Role): Promise<void> => {
+    if (role !== 'user' && role !== 'admin') {
+        throw new UnauthorizedError('credentials_required', {
+            message: 'Only users can delete their own recipes.',
+        });
+    }
+
     if (id <= 0) throw new Error('Invalid recipe ID');
 
     const recipe = await recipeDb.getRecipeById({ id });
     if (!recipe) throw new Error(`Recipe with id ${id} does not exist.`);
+
     await recipeDb.deleteRecipe({ id });
 };
